Improve cli and refactor slightly.
After trying and failing to find a cli parsing library that suited my
needs, I have elected to go for an api that will allow for doing this
manually via the `flag` package, which required some minor refactoring
of types. Search query logic is unimplemented but the flag and
associated validation are included in this commit.