Overhaul UI to be more like website.
After attempting to make the one-line-per-item idea work, it has become
obvious that this is just not possible if we want to include enough
relevant information. The UI has therefore been overhauled to a
two-lines-per-item system similar to the website's design.

This new UI now includes authors (with markdown links). It bolds story
titles and italicizes comment text.